+++ lib/src/otlp_http_sender.dart
@@ -1,1 +1,0 @@
-export 'agent/otlp/otlp_http_sender.dart';