+++ lib/src/otlp_metrics_exporter.dart
@@ -1,1 +1,0 @@
-export 'agent/otlp/otlp_metrics_exporter.dart';