+++ lib/src/otlp_trace_exporter.dart
@@ -1,1 +1,0 @@
-export 'agent/otlp/otlp_trace_exporter.dart';