+++ lib/src/skywalking.dart
@@ -1,5 +1,0 @@
-import 'otlp_agent.dart';
-
-/// Backward-compatible entry name for [OtlpAgent].
-@Deprecated('Use OtlpAgent')
-typedef Skywalking = OtlpAgent;